+++ .gitignore
@@ -1,3 +1,4 @@
+_ASSETS
 .firebase
 .firebaserc
 .DS_Store
