+++ react-app/src/flash/Flash.tsx
