+++ react-app/src/flash/components/SkeletonVariants.tsx
