+++ react-app/src/flash/index.tsx
